# GDScript: Optimize non-constant `for`-`range`