--- a/modules/gdscript/tests/scripts/runtime/features/for_range_large_ints.gd
+++ b/modules/gdscript/tests/scripts/runtime/features/for_range_large_ints.gd
@@ -1,0 +1,7 @@
+# GH-83293
+
+func test():
+	for x in range(1 << 31, (1 << 31) + 3):
+		print(x)
+	for x in range(1 << 62, (1 << 62) + 3):
+		print(x)