* Fixes #83293.