@dalexeev 2023-01-18
Is there a way to improve error messages?

```gdscript
var a: Variant = 1
var b: Variant = 8
var c: Variant = 2
var d: Variant = []

for i in range(a, b, c): # 1, 3, 5, 7
    print(var_to_str(i))

for i in range(a, b, d): # Runtime error: Invalid call. Nonexistent 'Vector3i' constructor.
    print(var_to_str(i))
```