@Calinou 2023-02-06
Can you provide before/after benchmarks with a release export template?