@dalexeev 2023-02-07
I think we can defer this to 4.1 as it's not a blocker or a bug. And if someone has a problem allocating a large array, they can explicitly use `int`/`Vector2i`/`Vector3i` or replace `for` with `while` as a workaround.