@dalexeev 2023-06-03
### Benchmarks

**Test script**

```gdscript
extends Node

func _ready() -> void:
    await get_tree().create_timer(3.0).timeout
    print("Test 1. Simple loop.")
    await get_tree().create_timer(3.0).timeout
    var a := 50_000_000
    var start_time := Time.get_ticks_msec()
    for i in range(0, a):
        var x := i
    var time := Time.get_ticks_msec() - start_time
    print("%d msec" % time)

    await get_tree().create_timer(3.0).timeout
    print("Test 2. Nested loop.")
    await get_tree().create_timer(3.0).timeout
    a = 10_000
    start_time = Time.get_ticks_msec()
    for i in range(0, a):
        for j in range(i, a):
            var x := j
    time = Time.get_ticks_msec() - start_time
    print("%d msec" % time)
```

**Before (`dev_build=true tests=true`)**

```
Test 1. Simple loop. # ~1 GiB RAM.
7753 msec
Test 2. Nested loop.
8010 msec
```

**Before (`target=template_release`)**

```
Test 1. Simple loop. # ~1 GiB RAM.
1568 msec
Test 2. Nested loop.
1404 msec
```


**After (`dev_build=true tests=true`)**

```
Test 1. Simple loop. # ~0 GiB RAM.
2934 msec
Test 2. Nested loop.
2941 msec
```

**After (`target=template_release`)**

```
Test 1. Simple loop. # ~0 GiB RAM.
426 msec
Test 2. Nested loop.
433 msec
```