@dalexeev 2024-10-23
### Updated benchmarks

<details>
<summary>Test script</summary>

```gdscript
extends Node

var _mem_before: float
var _mem_after: float

func simple_alt_to() -> void:
    var a := 50_000_000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in int(a):
        if i == 0:
            _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
        var _x := i

func simple_alt_from_to() -> void:
    var a := 50_000_000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in Vector2i(0, a):
        if i == 0:
            _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
        var _x := i

func simple_alt_from_to_step() -> void:
    var a := 50_000_000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in Vector3i(0, a, 1):
        if i == 0:
            _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
        var _x := i

func simple_range_to() -> void:
    var a := 50_000_000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in range(a):
        if i == 0:
            _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
        var _x := i

func simple_range_from_to() -> void:
    var a := 50_000_000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in range(0, a):
        if i == 0:
            _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
        var _x := i

func simple_range_from_to_step() -> void:
    var a := 50_000_000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in range(0, a, 1):
        if i == 0:
            _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
        var _x := i

func nested_range_to() -> void:
    var a := 50_000
    var b := 1000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in range(a):
        for j in range(b):
            if i == 0 and j == 0:
                _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
            var _x := j

func nested_range_from_to() -> void:
    var a := 50_000
    var b := 1000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in range(0, a):
        for j in range(0, b):
            if i == 0 and j == 0:
                _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
            var _x := j

func nested_range_from_to_step() -> void:
    var a := 50_000
    var b := 1000
    _mem_before = Performance.get_monitor(Performance.MEMORY_STATIC)
    for i in range(0, a, 1):
        for j in range(0, b, 1):
            if i == 0 and j == 0:
                _mem_after = Performance.get_monitor(Performance.MEMORY_STATIC)
            var _x := j

func benchmark(test_name: String, test_func: Callable) -> void:
    var start_time := Time.get_ticks_msec()
    test_func.call()
    var mem := _mem_after - _mem_before
    var time := Time.get_ticks_msec() - start_time
    print("%s - %s - %d msec" % [test_name, String.humanize_size(int(mem)), time])

func _ready() -> void:
    print("Please wait...")
    await get_tree().create_timer(3.0).timeout
    benchmark("Simple alt to", simple_alt_to)
    benchmark("Simple alt from-to", simple_alt_from_to)
    benchmark("Simple alt from-to-step", simple_alt_from_to_step)
    benchmark("Simple range to", simple_range_to)
    benchmark("Simple range from-to", simple_range_from_to)
    benchmark("Simple range from-to-step", simple_range_from_to_step)
    benchmark("Nested range to", nested_range_to)
    benchmark("Nested range from-to", nested_range_from_to)
    benchmark("Nested range from-to-step", nested_range_from_to_step)
    print("Done!")
```

</details>

(Monitor `Performance.MEMORY_STATIC` is not available in release builds.)

```
Before (dev_build=true tests=true)
==================================
Simple alt to - 0 B - 7306 msec
Simple alt from-to - 0 B - 7278 msec
Simple alt from-to-step - 0 B - 7243 msec
Simple range to - 2.00 GiB - 12619 msec
Simple range from-to - 2.00 GiB - 12396 msec
Simple range from-to-step - 2.00 GiB - 12544 msec
Nested range to - 2.03 MiB - 13774 msec
Nested range from-to - 2.03 MiB - 13782 msec
Nested range from-to-step - 2.03 MiB - 14062 msec

Before (target=template_release)
================================
Simple alt to - 843 msec
Simple alt from-to - 826 msec
Simple alt from-to-step - 835 msec
Simple range to - 2207 msec
Simple range from-to - 2206 msec
Simple range from-to-step - 2245 msec
Nested range to - 2329 msec
Nested range from-to - 2252 msec
Nested range from-to-step - 2251 msec

After (dev_build=true tests=true)
==================================
Simple alt to - 0 B - 7267 msec
Simple alt from-to - 0 B - 6984 msec
Simple alt from-to-step - 0 B - 7150 msec
Simple range to - 0 B - 7488 msec
Simple range from-to - 0 B - 7527 msec
Simple range from-to-step - 0 B - 7465 msec
Nested range to - 0 B - 9333 msec
Nested range from-to - 0 B - 9305 msec
Nested range from-to-step - 0 B - 9329 msec

After (target=template_release)
================================
Simple alt to - 872 msec
Simple alt from-to - 825 msec
Simple alt from-to-step - 828 msec
Simple range to - 860 msec
Simple range from-to - 912 msec
Simple range from-to-step - 907 msec
Nested range to - 1274 msec
Nested range from-to - 1276 msec
Nested range from-to-step - 1344 msec
```